requirements.txt
streamlit
pillow==9.4.0
rembg
pybase64